+++ Assets/ScriptableObjects/Grid Manager.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 65b6e3218492a4addaf7c46daee98dd7
-folderAsset: yes
-DefaultImporter:
+guid: 08a6b1332d45c4c8db970ca0669179d9
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
