+++ Assets/Scripts/_ScriptableObjects.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 65b6e3218492a4addaf7c46daee98dd7
+guid: 4426547bac907415e8c240c35ce7cf9e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
