--- Assets/Scripts/_ScriptableObjects/Biome.cs
@@ -1,14 +1,0 @@
-using UnityEngine;
-
-[System.Serializable]
-public struct Biome
-{
-    [Tooltip("Ranges from 0 to 1.")]
-    public float MinHeight;
-
-    [Tooltip("Ranges from 0 to 1.")]
-    public float MinMoisture;
-
-    [Tooltip("Ranges from 0 to 1.")]
-    public float MinHeat;
-}
