--- Assets/Scripts/_ScriptableObjects/Biome.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3e90f14e536e3444dbb33ad9c57c6126
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
